requirements.txt
django==4.2.7
pytest-django==4.7.0